feat: expose refresh callbacks to all JS sources
