+++ app/src/main/assets/updateLog.md
@@ -10,6 +10,7 @@
 - 修复连续解析书源规则时意外删除 HTML 节点，导致书籍链接误指向作者主页或发现页的问题
 - 修复启用 Cronet 时订阅正文视频、图片等子资源仍绕过应用网络请求的问题
 - 修复定时任务登录 URL 输入框无法换行及使用代码编辑器的问题
+- 登录 JS 与普通 JS 书源统一支持刷新书籍详情、目录和正文
 
 **2026/09/05**
 - 阅读内容编辑新增普通文本与正则搜索、大小写开关、循环定位及匹配计数，支持独立拖动正文位置并在编辑后更新
