+++ app/src/main/java/io/legado/app/ui/login/SourceLoginJsExtensions.kt
@@ -3,14 +3,12 @@ package io.legado.app.ui.login
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
 import io.legado.app.R
-import io.legado.app.constant.EventBus
 import io.legado.app.data.entities.BaseSource
 import io.legado.app.data.entities.HttpTTS
 import io.legado.app.model.ReadAloud
 import io.legado.app.ui.rss.read.RssJsExtensions
 import io.legado.app.ui.widget.dialog.BottomWebViewDialog
 import io.legado.app.utils.FileUtils
-import io.legado.app.utils.postEvent
 import io.legado.app.utils.sendToClip
 import io.legado.app.utils.showDialogFragment
 import io.legado.app.utils.toastOnUi
@@ -43,18 +41,6 @@ class SourceLoginJsExtensions(
 
     fun refreshExplore() {
         callbackRef.get()?.reUiView()
-    }
-
-    fun refreshBookInfo() {
-        postEvent(EventBus.REFRESH_BOOK_INFO, true)
-    }
-
-    fun refreshBookToc() {
-        postEvent(EventBus.REFRESH_BOOK_TOC, true)
-    }
-
-    fun refreshContent() {
-        postEvent(EventBus.REFRESH_BOOK_CONTENT, true)
     }
 
     fun copyText(text: String) {
